Add CLA/Code of Conduct/Privacy docs and banner; enable Dokka/Kover; upgrade Kotlin/Gradle; fix readChar() conversion
- Add CLA.md, CLA_EXPLANATIONS.md, CODE_OF_CONDUCT.md, CONTRIBUTOR_PRIVACY.md, and .banner.png
- Upgrade Kotlin JVM plugin to 2.2.20; add Dokka 2.0.0 and Kover 0.9.1 with configuration
- Update Gradle wrapper to 8.7 and set Dokka V2 plugin mode in gradle.properties
- Fix EndianInputStream.readChar() by converting via toInt().toChar()

Semver Impact: PATCH

+++ src/main/kotlin/one/wabbit/io/EndianInputStream.kt
@@ -165,7 +165,7 @@ class EndianInputStream(val base: InputStream, val defaultByteOrder: ByteOrder) 
     }
     override fun readShort(): Short = readShort(defaultByteOrder)
     override fun readUnsignedShort(): Int = readUShort(defaultByteOrder).toInt()
-    override fun readChar(): Char = readShort(defaultByteOrder).toChar()
+    override fun readChar(): Char = readShort(defaultByteOrder).toInt().toChar()
     override fun readInt(): Int = readInt(defaultByteOrder)
     override fun readLong(): Long = readLong(defaultByteOrder)
     override fun readFloat(): Float = java.lang.Float.intBitsToFloat(readInt())
